banner and minor bug fixes

+++ src/index.ts
@@ -1,4 +1,11 @@
-import { Guild, Message, MessageEmbed, TextChannel, User, VoiceState } from 'discord.js';
+import {
+    Guild,
+    Message,
+    MessageEmbed,
+    TextChannel,
+    User,
+    VoiceState,
+} from 'discord.js';
 import { Queue, IQueue } from './models/queue_schema';
 import { connect } from 'mongoose';
 import Client from './client/Client';
@@ -9,7 +16,10 @@ import shuffle from './functions/shuffle';
 import pause from './functions/pause';
 import Player from './models/player_schema';
 import loop from './functions/loop';
-const client: Client = new Client({ partials: ['MESSAGE', 'CHANNEL', 'REACTION'] });
+import antispam from './functions/antispam';
+const client: Client = new Client({
+    partials: ['MESSAGE', 'CHANNEL', 'REACTION'],
+});
 const prefix = '!';
 const db = process.env.DB_CONNECTION;
 let dbConnected = false;
@@ -29,22 +39,13 @@ connect(db, {
         console.log(err);
     });
 
-
-
-
-const usersMap: Map<string, any> = new Map();
-const LIMIT = 7;
-const DIFF = 5000;
-const TIME = 1000;
-
-
 client.on('ready', async () => {
     console.log('I am ready to pop');
     const commandFiles = fs
-        .readdirSync(__dirname+ '/commands')
+        .readdirSync(__dirname + '/commands')
         .filter(file => {
-            if(file.endsWith('.js')) return file;
-            if(file.endsWith('.ts')) return file
+            if (file.endsWith('.js')) return file;
+            if (file.endsWith('.ts')) return file;
         });
     for (const file of commandFiles) {
         const command = await import(`./commands/${file}`);
@@ -52,12 +53,13 @@ client.on('ready', async () => {
     }
 });
 client.on('guildCreate', (guild: Guild) => {
-    if(!dbConnected) return;
+    if (!dbConnected) return;
     const infoChannel: TextChannel = guild.channels.cache
         .filter(
-            channel => channel
-                .permissionsFor(client.user as User)
-                ?.has('SEND_MESSAGES') as boolean
+            channel =>
+                channel
+                    .permissionsFor(client.user as User)
+                    ?.has('SEND_MESSAGES') as boolean
         )
         .filter(channel => channel.type === 'text')
         .first() as TextChannel;
@@ -69,111 +71,79 @@ client.on('guildCreate', (guild: Guild) => {
     infoChannel.send(helloMessage);
 });
 
-
 // updates bot's voiceChannelId inside database
-client.on("voiceStateUpdate", async (oldState: VoiceState, newState:VoiceState)=>{
-    if ( oldState.channelID === null || typeof oldState.channelID == 'undefined') return;
-    if (newState.id !== client.user?.id) return;
-    
-    
-    if (!newState.channel) {
-        const voiceConnection = newState.guild!.me!.voice;
-        await Queue.updateOne({guildId: newState.guild.id}, {
-            //@ts-ignore
-            $unset: { voiceChannelId:1 },
-            queue: []
-        });
-        await Queue.findOne({guildId: newState.guild.id}, async (err: Error, serverQueue: IQueue)=>{
-            if(serverQueue){
-                const textChannel: TextChannel | undefined = newState.guild.channels.cache.get(
-                    serverQueue.textChannelId
-                ) as TextChannel;
-                if(!textChannel) return;
-                const playerEmbedMessage:Message | undefined = await textChannel.messages.fetch(serverQueue.playerMessageId)
-                    .catch(err=>undefined)
-                if(!playerEmbedMessage) return;
-                playerEmbedMessage.edit(new Player());
+client.on(
+    'voiceStateUpdate',
+    async (oldState: VoiceState, newState: VoiceState) => {
+        if (
+            oldState.channelID === null ||
+            typeof oldState.channelID == 'undefined'
+        )
+            return;
+        if (newState.id !== client.user?.id) return;
 
-                const queueEmbedMessage = await textChannel.messages.fetch(
-                    serverQueue.queueTextMessageId
-                ).catch(err=>undefined)
-                if(!queueEmbedMessage) return;
-                queueEmbedMessage.delete();
-            }   
-        })
-        voiceConnection.connection?.dispatcher.end();
-        return;
+        if (!newState.channel) {
+            const voiceConnection = newState.guild!.me!.voice;
+            await Queue.updateOne(
+                { guildId: newState.guild.id },
+                {
+                    //@ts-ignore
+                    $unset: { voiceChannelId: 1 },
+                    queue: [],
+                }
+            );
+            await Queue.findOne(
+                { guildId: newState.guild.id },
+                async (err: Error, serverQueue: IQueue) => {
+                    if (serverQueue) {
+                        const textChannel: TextChannel | undefined =
+                            newState.guild.channels.cache.get(
+                                serverQueue.textChannelId
+                            ) as TextChannel;
+                        if (!textChannel) return;
+                        const playerEmbedMessage: Message | undefined =
+                            await textChannel.messages
+                                .fetch(serverQueue.playerMessageId)
+                                .catch(err => undefined);
+                        if (!playerEmbedMessage) return;
+                        playerEmbedMessage.edit(new Player());
+
+                        const queueEmbedMessage = await textChannel.messages
+                            .fetch(serverQueue.queueTextMessageId)
+                            .catch(err => undefined);
+                        if (!queueEmbedMessage) return;
+                        queueEmbedMessage.delete();
+                    }
+                }
+            );
+            voiceConnection.connection?.dispatcher.end();
+            return;
+        }
+
+        const serverQueue = await Queue.findOne({ guildId: newState.guild.id });
+        if (!serverQueue) return;
+
+        // when the bot is moved lines below update his voice channel inside databae
+
+        serverQueue.voiceChannelId = newState.channel.id;
+        return await serverQueue.save();
     }
-    
-    const serverQueue = await Queue.findOne({guildId: newState.guild.id});
-    if (!serverQueue) return;
+);
 
-    // when the bot is moved lines below update his voice channel inside databae
-
-    serverQueue.voiceChannelId = newState.channel.id;
-    return await serverQueue.save();
-})
-
-
+let textChannelId = new Map();
 client.on('message', async (message: Message) => {
     if (message.author.bot) return;
     if (!message.guild) return;
 
-    if(usersMap.has(message.author.id)) {
-        const userData = usersMap.get(message.author.id);
-        const { lastMessage, timer } = userData;
-        const difference = message.createdTimestamp - lastMessage.createdTimestamp;
-        let msgCount = userData.msgCount;
-        console.log(difference);
-
-        if(difference > DIFF) {
-            clearTimeout(timer);
-            console.log('Cleared Timeout');
-            userData.msgCount = 1;
-            userData.lastMessage = message;
-            userData.timer = setTimeout(() => {
-                usersMap.delete(message.author.id);
-                console.log('Removed from map.')
-            }, TIME);
-            usersMap.set(message.author.id, userData)
-        }
-        else {
-            ++msgCount;
-            if(parseInt(msgCount) === LIMIT) {
-              //@ts-ignore
-              message.channel.bulkDelete(LIMIT);
-              return
-               
-            } else {
-                userData.msgCount = msgCount;
-                usersMap.set(message.author.id, userData);
-            }
-        }
-    }
-    else {
-        let fn = setTimeout(() => {
-            usersMap.delete(message.author.id);
-            console.log('Removed from map.')
-        }, TIME);
-        usersMap.set(message.author.id, {
-            msgCount: 1,
-            lastMessage : message,
-            timer : fn
-        });
-    }
-
-
-
-
     if (!isInitialized) {
-        const queue = await Queue.findOne({ guildId: message.guild.id });
-        if (queue) isInitialized = true;
+        const serverQueue = await Queue.findOne({ guildId: message.guild.id });
+        if (serverQueue) isInitialized = true;
     }
     const [command, ...args] = message.content.slice(prefix.length).split(/ +/);
 
     if (!isInitialized) {
         if (!client.commands.has(command)) return;
-        if (command!="init") {
+        if (command != 'init') {
             return message.channel.send(`U have to use ${prefix}init first!`);
         }
         try {
@@ -183,13 +153,33 @@ client.on('message', async (message: Message) => {
             message.reply('there was an error trying to execute that command!');
         }
     } else {
-        if(message.content.startsWith(prefix)) {
+        if (!textChannelId.has(message.guild.id)) {
+            const serverQueue = await Queue.findOne({
+                guildId: message.guild.id,
+            });
+            if (!serverQueue) return;
+            textChannelId.set(message.guild.id, {
+                textChannel: serverQueue.textChannelId,
+            });
+        }
+        const validTextChannel = textChannelId.get(message.guild.id);
+        if (message.channel.id != validTextChannel.textChannel) {
+            return;
+        }
+        const usersMap: Map<string, any> = new Map();
+        const limit = 7;
+        const diff = 5000;
+        const time = 1000;
+        antispam(message, usersMap, time, diff, limit);
+        if (message.content.startsWith(prefix)) {
             try {
                 await message.delete();
                 client.commands.get(command)?.execute(message, args, client);
             } catch (error) {
                 console.error(error);
-                message.reply('there was an error trying to execute that command!');
+                message.reply(
+                    'there was an error trying to execute that command!'
+                );
             }
         } else {
             try {
@@ -198,36 +188,34 @@ client.on('message', async (message: Message) => {
                 return;
             }
         }
-
     }
 });
 
-client.on("messageReactionAdd", async (reaction, user)=>{
-    if(user.bot) return;
-    if(!reaction.message.guild) return;
-    const serverQueue = await Queue.findOne({guildId: reaction.message.guild.id});
-    if(!serverQueue) return;
-    if(reaction.message.id != serverQueue.playerMessageId) return;
-    if(reaction.emoji.name=="⏹️"){
-        console.log("stop");
+client.on('messageReactionAdd', async (reaction, user) => {
+    if (user.bot) return;
+    if (!reaction.message.guild) return;
+    const serverQueue = await Queue.findOne({
+        guildId: reaction.message.guild.id,
+    });
+    if (!serverQueue) return;
+    if (reaction.message.id != serverQueue.playerMessageId) return;
+    if (reaction.emoji.name == '⏹️') {
+        console.log('stop');
         stop(reaction.message.channel as TextChannel, user as User);
-    } else if(reaction.emoji.name=="⏭️"){
-        console.log("skip");
+    } else if (reaction.emoji.name == '⏭️') {
+        console.log('skip');
         skip(reaction.message.channel as TextChannel, user as User);
-    } else if(reaction.emoji.name=="⏯️"){
-        console.log("pause");
+    } else if (reaction.emoji.name == '⏯️') {
+        console.log('pause');
         pause(reaction.message.channel as TextChannel, user as User);
-    } else if(reaction.emoji.name=="🔀"){
-        console.log("shuffle");
+    } else if (reaction.emoji.name == '🔀') {
+        console.log('shuffle');
         shuffle(reaction.message.channel as TextChannel, user as User);
-    } else if(reaction.emoji.name=="🔄"){
-        console.log("loop");
+    } else if (reaction.emoji.name == '🔄') {
+        console.log('loop');
         loop(reaction.message.channel as TextChannel, user as User);
     }
     reaction.users.remove(user as User);
-})
-
-
-
+});
 
 client.start();
